+++ cirq-core/cirq/devices/__init__.py
@@ -16,13 +16,16 @@
 from cirq.devices.device import (
     Device,
     DeviceMetadata,
-    GridDeviceMetadata,
     SymmetricalQidPair,
 )
 
 from cirq.devices.grid_qubit import (
     GridQid,
     GridQubit,
+)
+
+from cirq.devices.griddevice_metadata import (
+    GridDeviceMetadata,
 )
 
 from cirq.devices.line_qubit import (
