+++ cirq-core/cirq/devices/device_test.py
@@ -119,99 +119,3 @@ def test_metadata_equality():
     eq.add_equality_group(cirq.DeviceMetadata(None, None))
 
     assert cirq.DeviceMetadata(None, graph) != cirq.DeviceMetadata(None, graph2)
-
-
-def test_griddevice_metadata():
-    qubits = cirq.GridQubit.rect(2, 3)
-    qubit_pairs = [(a, b) for a in qubits for b in qubits if a != b and a.is_adjacent(b)]
-
-    gateset = cirq.Gateset(cirq.XPowGate, cirq.YPowGate, cirq.ZPowGate, cirq.CZ)
-    metadata = cirq.GridDeviceMetadata(qubit_pairs, gateset)
-
-    expected_pairings = frozenset(
-        {
-            (cirq.GridQubit(0, 0), cirq.GridQubit(0, 1)),
-            (cirq.GridQubit(0, 1), cirq.GridQubit(0, 2)),
-            (cirq.GridQubit(0, 1), cirq.GridQubit(1, 1)),
-            (cirq.GridQubit(0, 2), cirq.GridQubit(1, 2)),
-            (cirq.GridQubit(1, 0), cirq.GridQubit(1, 1)),
-            (cirq.GridQubit(1, 1), cirq.GridQubit(1, 2)),
-            (cirq.GridQubit(0, 0), cirq.GridQubit(1, 0)),
-        }
-    )
-    assert metadata.qubit_set == frozenset(qubits)
-    assert metadata.qubit_pairs == expected_pairings
-    assert metadata.gateset == gateset
-    expected_graph = nx.Graph()
-    expected_graph.add_edges_from(sorted(list(expected_pairings)), directed=False)
-    assert metadata.nx_graph.edges() == expected_graph.edges()
-    assert metadata.nx_graph.nodes() == expected_graph.nodes()
-    assert metadata.gate_durations is None
-
-
-def test_griddevice_metadata_bad_durations():
-    qubits = tuple(cirq.GridQubit.rect(1, 2))
-
-    gateset = cirq.Gateset(cirq.XPowGate, cirq.YPowGate)
-    invalid_duration = {
-        cirq.Gateset(cirq.XPowGate): cirq.Duration(nanos=1),
-        cirq.Gateset(cirq.ZPowGate): cirq.Duration(picos=1),
-    }
-    with pytest.raises(ValueError, match="ZPowGate"):
-        cirq.GridDeviceMetadata([qubits], gateset, gate_durations=invalid_duration)
-
-
-def test_griddevice_json_load():
-    qubits = cirq.GridQubit.rect(2, 3)
-    qubit_pairs = [(a, b) for a in qubits for b in qubits if a != b and a.is_adjacent(b)]
-    gateset = cirq.Gateset(cirq.XPowGate, cirq.YPowGate, cirq.ZPowGate)
-    duration = {
-        cirq.Gateset(cirq.XPowGate): cirq.Duration(nanos=1),
-        cirq.Gateset(cirq.YPowGate): cirq.Duration(picos=2),
-        cirq.Gateset(cirq.ZPowGate): cirq.Duration(picos=3),
-    }
-    metadata = cirq.GridDeviceMetadata(qubit_pairs, gateset, gate_durations=duration)
-    rep_str = cirq.to_json(metadata)
-    assert metadata == cirq.read_json(json_text=rep_str)
-
-
-def test_griddevice_metadata_equality():
-    qubits = cirq.GridQubit.rect(2, 3)
-    qubit_pairs = [(a, b) for a in qubits for b in qubits if a != b and a.is_adjacent(b)]
-    gateset = cirq.Gateset(cirq.XPowGate, cirq.YPowGate, cirq.ZPowGate)
-    duration = {
-        cirq.Gateset(cirq.XPowGate): cirq.Duration(nanos=1),
-        cirq.Gateset(cirq.YPowGate): cirq.Duration(picos=3),
-        cirq.Gateset(cirq.ZPowGate): cirq.Duration(picos=2),
-    }
-    duration2 = {
-        cirq.Gateset(cirq.XPowGate): cirq.Duration(nanos=10),
-        cirq.Gateset(cirq.YPowGate): cirq.Duration(picos=13),
-        cirq.Gateset(cirq.ZPowGate): cirq.Duration(picos=12),
-    }
-    metadata = cirq.GridDeviceMetadata(qubit_pairs, gateset, gate_durations=duration)
-    metadata2 = cirq.GridDeviceMetadata(qubit_pairs[:2], gateset, gate_durations=duration)
-    metadata3 = cirq.GridDeviceMetadata(qubit_pairs, gateset, gate_durations=None)
-    metadata4 = cirq.GridDeviceMetadata(qubit_pairs, gateset, gate_durations=duration2)
-    metadata5 = cirq.GridDeviceMetadata(reversed(qubit_pairs), gateset, gate_durations=duration)
-
-    eq = cirq.testing.EqualsTester()
-    eq.add_equality_group(metadata)
-    eq.add_equality_group(metadata2)
-    eq.add_equality_group(metadata3)
-    eq.add_equality_group(metadata4)
-
-    assert metadata == metadata5
-
-
-def test_repr():
-    qubits = cirq.GridQubit.rect(2, 3)
-    qubit_pairs = [(a, b) for a in qubits for b in qubits if a != b and a.is_adjacent(b)]
-    gateset = cirq.Gateset(cirq.XPowGate, cirq.YPowGate, cirq.ZPowGate)
-    duration = {
-        cirq.Gateset(cirq.XPowGate): cirq.Duration(nanos=1),
-        cirq.Gateset(cirq.YPowGate): cirq.Duration(picos=3),
-        cirq.Gateset(cirq.ZPowGate): cirq.Duration(picos=2),
-    }
-    metadata = cirq.GridDeviceMetadata(qubit_pairs, gateset, gate_durations=duration)
-    cirq.testing.assert_equivalent_repr(metadata)
